implement metachara check and escape
