+++ chat-api/gen/http/openapi.json
@@ -1,1 +1,1 @@
-{"swagger":"2.0","info":{"title":"Chat Service","description":"Service for chat app, a Goa teaser","version":""},"host":"172.25.0.4:8000","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/chatroom/{id}":{"get":{"tags":["chatapi"],"summary":"getchat chatapi","operationId":"chatapi#getchat","parameters":[{"name":"id","in":"path","description":"room id","required":true,"type":"integer"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/ChatapiGoaChatResponseCollection"}}},"schemes":["http"]}}},"definitions":{"ChatapiGoaChatResponseCollection":{"title":"Mediatype identifier: application/vnd.goa.chat; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/GoaChatResponse"},"description":"GetchatResponseBody is the result type for an array of GoaChatResponse (default view)","example":[{"Chat":"Est eius et doloribus ex.","Member":"Sit soluta.","PostDt":"1989-01-28T15:35:39Z","RoomName":"Et quo rerum.","UserId":"Consequuntur et aut.","id":4311757619097375690},{"Chat":"Est eius et doloribus ex.","Member":"Sit soluta.","PostDt":"1989-01-28T15:35:39Z","RoomName":"Et quo rerum.","UserId":"Consequuntur et aut.","id":4311757619097375690},{"Chat":"Est eius et doloribus ex.","Member":"Sit soluta.","PostDt":"1989-01-28T15:35:39Z","RoomName":"Et quo rerum.","UserId":"Consequuntur et aut.","id":4311757619097375690},{"Chat":"Est eius et doloribus ex.","Member":"Sit soluta.","PostDt":"1989-01-28T15:35:39Z","RoomName":"Et quo rerum.","UserId":"Consequuntur et aut.","id":4311757619097375690}]},"GoaChatResponse":{"title":"Mediatype identifier: application/vnd.goa.chat; view=default","type":"object","properties":{"Chat":{"type":"string","description":"chat","example":"Saepe est omnis."},"Member":{"type":"string","description":"member","example":"Ut placeat non."},"PostDt":{"type":"string","example":"2008-06-29T09:15:30Z","format":"date-time"},"RoomName":{"type":"string","description":"room name","example":"Consequatur non iste."},"UserId":{"type":"string","description":"user id","example":"Quam qui commodi."},"id":{"type":"integer","description":"room id","example":6367248016073463729,"format":"int64"}},"description":"All chat (default view)","example":{"Chat":"Est et voluptas expedita recusandae.","Member":"Blanditiis excepturi non voluptas quisquam.","PostDt":"1996-07-21T01:42:47Z","RoomName":"Est eos explicabo praesentium.","UserId":"Aliquam dolor necessitatibus.","id":4000844251158404115},"required":["id","UserId","RoomName","Member","Chat","PostDt"]}}}
+{"swagger":"2.0","info":{"title":"Chat Service","description":"Service for chat app, a Goa teaser","version":""},"host":"172.25.0.4:8000","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/chatroom/{id}":{"get":{"tags":["chatapi"],"summary":"getchat chatapi","operationId":"chatapi#getchat","parameters":[{"name":"id","in":"path","description":"room id","required":true,"type":"integer"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/ChatapiGoaChatResponseCollection"}}},"schemes":["http"]}}},"definitions":{"ChatapiGoaChatResponseCollection":{"title":"Mediatype identifier: application/vnd.goa.chat; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/GoaChatResponse"},"description":"GetchatResponseBody is the result type for an array of GoaChatResponse (default view)","example":[{"Chat":"Est eius et doloribus ex.","Id":4311757619097375690,"Member":"Sit soluta.","PostDt":"1989-01-28T15:35:39Z","RoomName":"Et quo rerum.","UserId":"Consequuntur et aut."},{"Chat":"Est eius et doloribus ex.","Id":4311757619097375690,"Member":"Sit soluta.","PostDt":"1989-01-28T15:35:39Z","RoomName":"Et quo rerum.","UserId":"Consequuntur et aut."},{"Chat":"Est eius et doloribus ex.","Id":4311757619097375690,"Member":"Sit soluta.","PostDt":"1989-01-28T15:35:39Z","RoomName":"Et quo rerum.","UserId":"Consequuntur et aut."},{"Chat":"Est eius et doloribus ex.","Id":4311757619097375690,"Member":"Sit soluta.","PostDt":"1989-01-28T15:35:39Z","RoomName":"Et quo rerum.","UserId":"Consequuntur et aut."}]},"GoaChatResponse":{"title":"Mediatype identifier: application/vnd.goa.chat; view=default","type":"object","properties":{"Chat":{"type":"string","description":"chat","example":"Saepe est omnis."},"Id":{"type":"integer","description":"room id","example":6367248016073463729,"format":"int64"},"Member":{"type":"string","description":"member","example":"Ut placeat non."},"PostDt":{"type":"string","example":"2008-06-29T09:15:30Z","format":"date-time"},"RoomName":{"type":"string","description":"room name","example":"Consequatur non iste."},"UserId":{"type":"string","description":"user id","example":"Quam qui commodi."}},"description":"All chat (default view)","example":{"Chat":"Est et voluptas expedita recusandae.","Id":4000844251158404115,"Member":"Blanditiis excepturi non voluptas quisquam.","PostDt":"1996-07-21T01:42:47Z","RoomName":"Est eos explicabo praesentium.","UserId":"Aliquam dolor necessitatibus."},"required":["Id","UserId","RoomName","Member","Chat","PostDt"]}}}
