+++ chat-api/gen/http/openapi3.json
@@ -1,1 +1,1 @@
-{"openapi":"3.0.3","info":{"title":"Chat Service","description":"Service for chat app, a Goa teaser","version":"1.0"},"servers":[{"url":"http://172.25.0.4:8000"}],"paths":{"/chatroom/{id}":{"get":{"tags":["chatapi"],"summary":"getchat chatapi","operationId":"chatapi#getchat","parameters":[{"name":"id","in":"path","description":"room id","required":true,"schema":{"type":"integer","description":"room id","example":517066950974973379,"format":"int64"},"example":5734865064250683693}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GoaChatCollection"},"example":[{"Chat":"Voluptas quod qui.","Member":"Voluptatem rerum rerum neque assumenda.","PostDt":"2013-12-17T20:03:48Z","RoomName":"Culpa id suscipit rem voluptates.","UserId":"Omnis molestiae.","id":714744356156329433},{"Chat":"Voluptas quod qui.","Member":"Voluptatem rerum rerum neque assumenda.","PostDt":"2013-12-17T20:03:48Z","RoomName":"Culpa id suscipit rem voluptates.","UserId":"Omnis molestiae.","id":714744356156329433},{"Chat":"Voluptas quod qui.","Member":"Voluptatem rerum rerum neque assumenda.","PostDt":"2013-12-17T20:03:48Z","RoomName":"Culpa id suscipit rem voluptates.","UserId":"Omnis molestiae.","id":714744356156329433},{"Chat":"Voluptas quod qui.","Member":"Voluptatem rerum rerum neque assumenda.","PostDt":"2013-12-17T20:03:48Z","RoomName":"Culpa id suscipit rem voluptates.","UserId":"Omnis molestiae.","id":714744356156329433}]}}}}}}},"components":{"schemas":{"GoaChat":{"type":"object","properties":{"Chat":{"type":"string","description":"chat","example":"Adipisci autem est iusto natus."},"Member":{"type":"string","description":"member","example":"Facilis quam ut doloremque repudiandae ipsam."},"PostDt":{"type":"string","example":"2014-01-26T20:09:51Z","format":"date-time"},"RoomName":{"type":"string","description":"room name","example":"Sint sit ullam quas."},"UserId":{"type":"string","description":"user id","example":"Est saepe in."},"id":{"type":"integer","description":"room id","example":561858919428618001,"format":"int64"}},"description":"All chat","example":{"Chat":"Debitis et.","Member":"Temporibus in provident perspiciatis.","PostDt":"1983-03-15T21:26:55Z","RoomName":"Atque consequatur repudiandae sunt non excepturi.","UserId":"Consequatur blanditiis at ad in eum.","id":7405198893839022047},"required":["id","UserId","RoomName","Member","Chat","PostDt"]},"GoaChatCollection":{"type":"array","items":{"$ref":"#/components/schemas/GoaChat"},"example":[{"Chat":"Voluptas quod qui.","Member":"Voluptatem rerum rerum neque assumenda.","PostDt":"2013-12-17T20:03:48Z","RoomName":"Culpa id suscipit rem voluptates.","UserId":"Omnis molestiae.","id":714744356156329433},{"Chat":"Voluptas quod qui.","Member":"Voluptatem rerum rerum neque assumenda.","PostDt":"2013-12-17T20:03:48Z","RoomName":"Culpa id suscipit rem voluptates.","UserId":"Omnis molestiae.","id":714744356156329433}]}}},"tags":[{"name":"chatapi","description":"The service performs get chat."}]}
+{"openapi":"3.0.3","info":{"title":"Chat Service","description":"Service for chat app, a Goa teaser","version":"1.0"},"servers":[{"url":"http://172.25.0.4:8000"}],"paths":{"/chatroom/{id}":{"get":{"tags":["chatapi"],"summary":"getchat chatapi","operationId":"chatapi#getchat","parameters":[{"name":"id","in":"path","description":"room id","required":true,"schema":{"type":"integer","description":"room id","example":517066950974973379,"format":"int64"},"example":5734865064250683693}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GoaChatCollection"},"example":[{"Chat":"Voluptas quod qui.","Id":714744356156329433,"Member":"Voluptatem rerum rerum neque assumenda.","PostDt":"2013-12-17T20:03:48Z","RoomName":"Culpa id suscipit rem voluptates.","UserId":"Omnis molestiae."},{"Chat":"Voluptas quod qui.","Id":714744356156329433,"Member":"Voluptatem rerum rerum neque assumenda.","PostDt":"2013-12-17T20:03:48Z","RoomName":"Culpa id suscipit rem voluptates.","UserId":"Omnis molestiae."},{"Chat":"Voluptas quod qui.","Id":714744356156329433,"Member":"Voluptatem rerum rerum neque assumenda.","PostDt":"2013-12-17T20:03:48Z","RoomName":"Culpa id suscipit rem voluptates.","UserId":"Omnis molestiae."},{"Chat":"Voluptas quod qui.","Id":714744356156329433,"Member":"Voluptatem rerum rerum neque assumenda.","PostDt":"2013-12-17T20:03:48Z","RoomName":"Culpa id suscipit rem voluptates.","UserId":"Omnis molestiae."}]}}}}}}},"components":{"schemas":{"GoaChat":{"type":"object","properties":{"Chat":{"type":"string","description":"chat","example":"Adipisci autem est iusto natus."},"Id":{"type":"integer","description":"room id","example":561858919428618001,"format":"int64"},"Member":{"type":"string","description":"member","example":"Facilis quam ut doloremque repudiandae ipsam."},"PostDt":{"type":"string","example":"2014-01-26T20:09:51Z","format":"date-time"},"RoomName":{"type":"string","description":"room name","example":"Sint sit ullam quas."},"UserId":{"type":"string","description":"user id","example":"Est saepe in."}},"description":"All chat","example":{"Chat":"Debitis et.","Id":7405198893839022047,"Member":"Temporibus in provident perspiciatis.","PostDt":"1983-03-15T21:26:55Z","RoomName":"Atque consequatur repudiandae sunt non excepturi.","UserId":"Consequatur blanditiis at ad in eum."},"required":["Id","UserId","RoomName","Member","Chat","PostDt"]},"GoaChatCollection":{"type":"array","items":{"$ref":"#/components/schemas/GoaChat"},"example":[{"Chat":"Voluptas quod qui.","Id":714744356156329433,"Member":"Voluptatem rerum rerum neque assumenda.","PostDt":"2013-12-17T20:03:48Z","RoomName":"Culpa id suscipit rem voluptates.","UserId":"Omnis molestiae."},{"Chat":"Voluptas quod qui.","Id":714744356156329433,"Member":"Voluptatem rerum rerum neque assumenda.","PostDt":"2013-12-17T20:03:48Z","RoomName":"Culpa id suscipit rem voluptates.","UserId":"Omnis molestiae."}]}}},"tags":[{"name":"chatapi","description":"The service performs get chat."}]}
