+++ web-server/public/css/logout.css
@@ -1,0 +1,11 @@
+body {background-color:#330033;}
+
+div.logout{
+  text-align:center;
+  color: #FFFFFF;
+}
+
+div.logout a{
+  text-align:center;
+  color: #FFFFFF;
+}
