+++ web-server/public/css/resistration.css
@@ -1,0 +1,22 @@
+body {background-color:#330033;}
+  
+div.container{
+  background-color:#FFFFFF;
+  position: relative;
+  width: 350px;
+  margin-top: 15%;
+  margin-bottom: 25%;
+  margin-left: auto;
+  margin-right: auto;
+  padding-top: 1em;
+  padding-bottom: 0.5em;
+}
+
+form.resister-form {
+  text-align: center;
+  line-height:0.5;
+}
+
+#login-link{
+  text-align: center;
+}
