+++ web-server/public/css/resistrationcompleted.css
@@ -1,0 +1,11 @@
+body {background-color:#330033;}
+
+div.resistration-completed{
+  text-align:center;
+  color: #FFFFFF;
+}
+
+div.resistration-completed a{
+  text-align:center;
+  color: #FFFFFF;
+}
