+++ web-server/public/css/withdrawal.css
@@ -1,0 +1,18 @@
+body {background-color:#330033;}
+
+div.container{
+  background-color:#FFFFFF;
+  position: relative;
+  width: 250px;
+  margin-top: 15%;
+  margin-bottom: 25%;
+  margin-left: auto;
+  margin-right: auto;
+  padding-top: 1em;
+  padding-bottom: 0.5em;
+}
+
+form.withdrawal-form {
+  text-align: center;
+  line-height:0.5;
+}
