+++ web-server/public/css/withdrawalcompleted.css
@@ -1,0 +1,11 @@
+body {background-color:#330033;}
+
+div.withdrawal-completed{
+  text-align:center;
+  color: #FFFFFF;
+}
+
+div.withdrawal-completed a{
+  text-align:center;
+  color: #FFFFFF;
+}
