+++ web-server/templates/login.html
@@ -2,7 +2,7 @@
 <html>
 <head>
     <meta charset="UTF-8">
-    <link rel="stylesheet" type="text/css" href="/public/css/cssforlogin.css">
+    <link rel="stylesheet" type="text/css" href="/public/css/login.css">
     <title>ログイン画面</title>
 </head>
 <body>
@@ -15,7 +15,7 @@
   <p id="resistration"><a href="resistration">新規登録はこちら</a></p>
   </div>
   
-  <script type="text/javascript" src="/public/scripts/scriptforlogin.js">
+  <script type="text/javascript" src="/public/scripts/login.js">
   </script>
 </body>
 </html>
