+++ web-server/templates/logout.html
@@ -2,10 +2,14 @@
 <html>
 <head>
     <meta charset="UTF-8">
-    <title>マイページ</title>
+    <link rel="stylesheet" type="text/css" href="/public/css/logout.css">
+    <title>ログアウトしました。</title>
 </head>
 <body>
-  <h2>ログアウトしました</h2>
+  <div class="logout">
+    <h2>ログアウトしました</h2>
   <a href="http://172.25.0.2/login">ログインページに戻る</a>
+  </div>
+</body>
 </body>
 </html>
