+++ web-server/templates/mypage.html
@@ -2,7 +2,7 @@
 <html>
 <head>
     <meta charset="UTF-8">
-    <link rel="stylesheet" type="text/css" href="/public/css/cssformypage.css">
+    <link rel="stylesheet" type="text/css" href="/public/css/mypage.css">
     <title>マイページ</title>
 </head>
 <body>
