+++ web-server/templates/mypage/chatroom.html
@@ -2,7 +2,7 @@
 <html>
 <head>
     <meta charset="UTF-8">
-    <link rel="stylesheet" type="text/css" href="/public/css/cssforchatroom.css">
+    <link rel="stylesheet" type="text/css" href="/public/css/chatroom.css">
     <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
     <title>マイページ</title>
 </head>
@@ -23,7 +23,7 @@
   </div>
   <iframe name="sendPhoto" style="width:0px;height:0px;border:0px;"></iframe>
 
-  <script type="text/javascript" src="/public/scripts/scriptforchatroom.js">
+  <script type="text/javascript" src="/public/scripts/chatroom.js">
   </script>
 </body>
 </html>
