+++ web-server/templates/resistration.html
@@ -2,6 +2,7 @@
 <html>
 <head>
     <meta charset="UTF-8">
+    <link rel="stylesheet" type="text/css" href="/public/css/resistration.css">
     <title>登録フォーム</title>
 </head>
 <body>
@@ -14,34 +15,11 @@
     <p>新規パスワード</p>
     <p>(英数字、10文字以内)</p>
     <p><input type="password" id="password" name="password"></p>
+    <p>※次の記号は使用できません。</p>
+    <p>（&#x3F、&#x24、&amp;、&#x3D、&#x2D、&gt;、&lt;、&#x2B、&#x3B、&#x3A、&#x2A、&#x7C、&#x27）</p>
     <p><input type="submit" id="login" name="login" value="登録"></p>
   </form>
-  <p id="resistration"><a href="/login">ログイン画面に戻る</a></p>
+  <p id="login-link"><a href="/login">ログイン画面に戻る</a></p>
   </div>
 </body>
 </html>
-
-<style>
-  body {background-color:#330033;}
-  
-  div.container{
-    background-color:#FFFFFF;
-    position: relative;
-    width: 250px;
-    margin-top: 15%;
-    margin-bottom: 25%;
-    margin-left: auto;
-    margin-right: auto;
-    padding-top: 1em;
-    padding-bottom: 0.5em;
-  }
-  
-  form.resister-form {
-    text-align: center;
-    line-height:0.5;
-  }
-  
-  #resistration{
-    text-align: center;
-  }
-  </style>
