+++ web-server/templates/resistrationcompleted.html
@@ -2,10 +2,13 @@
 <html>
 <head>
     <meta charset="UTF-8">
+    <link rel="stylesheet" type="text/css" href="/public/css/resistrationcompleted.css">
     <title>登録完了</title>
 </head>
 <body>
-  <h2>登録が完了しました</h2>
-  <a href="login">ログインページに戻る</a>
+  <div class="resistration-completed">
+    <h2>登録が完了しました</h2>
+    <a href="login">ログインページに戻る</a>
+  </div>
 </body>
 </html>
