+++ web-server/templates/sessionexpired.html
@@ -2,11 +2,14 @@
 <html>
 <head>
     <meta charset="UTF-8">
+    <link rel="stylesheet" type="text/css" href="/public/css/sessionexpired.css">
     <title>セッションの有効期限が切れています。</title>
 </head>
 <body>
+  <div class="sessionexpired">
   <h2>セッションの有効期限が切れています。</h2>
   <a href="http://172.25.0.2/login">ログインページに戻る</a>
+  </div>
   <script>
       document.cookie = "cookieName=; max-age=0";
   </script>
