+++ web-server/templates/withdrawal.html
@@ -2,13 +2,16 @@
 <html>
 <head>
     <meta charset="UTF-8">
+    <link rel="stylesheet" type="text/css" href="/public/css/withdrawal.css">
     <title>退会ページ</title>
 </head>
 <body>
-<form action="" method="post">
-    <p>ユーザーID：<input type="text" id="userId" name="userId"></p>
-    <p>パスワード：<input type="password" id="password" name="password"></p>
-    <p><input type="submit" id="withdrawal" name="withdrawal"></p>
-</form>
+<div class="container">
+  <form class="withdrawal-form" action="" method="post">
+    <p><input type="text" id="userId" name="userId" placeholder="ユーザーID"></p>
+    <p><input type="password" id="password" name="password" placeholder="パスワード"></p>
+    <p><input type="submit" id="withdrawal" name="withdrawal" value="退会"></p>
+  </form>
+</div>
 </body>
 </html>
