+++ web-server/templates/withdrawalcompleted.html
@@ -2,10 +2,13 @@
 <html>
 <head>
     <meta charset="UTF-8">
+    <link rel="stylesheet" type="text/css" href="/public/css/withdrawalcompleted.css">
     <title>退会完了</title>
 </head>
 <body>
-  <h2>退会が完了しました</h2>
-  <a href="login">トップページに戻る</a>
+  <div class="withdrawal-completed">
+    <h2>退会が完了しました</h2>
+    <a href="login">トップページに戻る</a>
+  </div>
 </body>
 </html>
